agent_backend/agent_platform_basic/pyproject.toml
[tool.poetry]
name = "agent-platform-basic"
version = "0.1.0"
description = "agent_platform_extensions is a platform extensions extending library built to power building LLM application."
authors = [""]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12.7"

flask = "~3.0.1"
flask-login = "~0.6.3"
flask-migrate = "~4.0.7"
flask-compress = "~1.17"
flask-restful = "~0.3.10"
flask-sqlalchemy = "~3.1.1"

resend = "~0.7.0"
redis = { version = "~5.0.3", extras = ["hiredis"] }
boto3 = "~1.35.16"
sentry-sdk = { version = "~1.44.1", extras = ["flask"] }
fastapi = "^0.112.0"
fastapi-pagination = "^0.12.26"
uvicorn = "^0.30.5"
aiobotocore = "~2.15.2"

pyjwt = "~2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
